refactor: split employee manager into components
