+++ components/AddEmployeeForm.tsx
@@ -1,0 +1,216 @@
+import { Button } from './ui/button'
+import { Input } from './ui/input'
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from './ui/card'
+import type { Department, WorkSchedule } from './employeeTypes'
+
+interface AddEmployeeFormProps {
+  formData: {
+    employee_code: string
+    dni: string
+    name: string
+    email: string
+    phone: string
+    role: string
+    position: string
+    base_salary: string
+    hire_date: string
+    department_id: string
+    work_schedule_id: string
+    bank_name: string
+    bank_account: string
+  }
+  setFormData: (_data: any) => void
+  loading: boolean
+  onSubmit: (_e: React.FormEvent) => void
+  departments: Department[]
+  workSchedules: WorkSchedule[]
+  onCancel: () => void
+}
+
+export default function AddEmployeeForm({
+  formData,
+  setFormData,
+  loading,
+  onSubmit,
+  departments,
+  workSchedules,
+  onCancel
+}: AddEmployeeFormProps) {
+  return (
+    <Card>
+      <CardHeader>
+        <CardTitle>Add New Employee</CardTitle>
+        <CardDescription>Enter the employee&apos;s information</CardDescription>
+      </CardHeader>
+      <CardContent>
+        <form onSubmit={onSubmit} className="grid grid-cols-1 md:grid-cols-2 gap-4">
+          <div>
+            <label className="block text-sm font-medium text-gray-700 mb-1">
+              Employee Code
+            </label>
+            <Input
+              value={formData.employee_code}
+              onChange={(e) => setFormData({ ...formData, employee_code: e.target.value })}
+              placeholder="EMP001"
+              required
+            />
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-gray-700 mb-1">
+              DNI
+            </label>
+            <Input
+              value={formData.dni}
+              onChange={(e) => setFormData({ ...formData, dni: e.target.value })}
+              placeholder="0801-1990-12345"
+              required
+            />
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-gray-700 mb-1">
+              Full Name
+            </label>
+            <Input
+              value={formData.name}
+              onChange={(e) => setFormData({ ...formData, name: e.target.value })}
+              placeholder="John Doe"
+              required
+            />
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-gray-700 mb-1">
+              Email
+            </label>
+            <Input
+              type="email"
+              value={formData.email}
+              onChange={(e) => setFormData({ ...formData, email: e.target.value })}
+              placeholder="john@company.com"
+            />
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-gray-700 mb-1">
+              Phone
+            </label>
+            <Input
+              value={formData.phone}
+              onChange={(e) => setFormData({ ...formData, phone: e.target.value })}
+              placeholder="+504 9999-9999"
+            />
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-gray-700 mb-1">
+              Position
+            </label>
+            <Input
+              value={formData.position}
+              onChange={(e) => setFormData({ ...formData, position: e.target.value })}
+              placeholder="Software Developer"
+              required
+            />
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-gray-700 mb-1">
+              Department
+            </label>
+            <select
+              value={formData.department_id}
+              onChange={(e) => setFormData({ ...formData, department_id: e.target.value })}
+              className="w-full p-2 border border-gray-300 rounded-md focus:ring-2 focus:ring-blue-500"
+            >
+              <option value="">Select Department</option>
+              {/* eslint-disable-next-line react/jsx-key */}
+              {departments.map((dept, index) => (
+                <option key={`dept-${index}`} value={dept.id}>{dept.name}</option>
+              ))}
+            </select>
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-gray-700 mb-1">
+              Work Schedule
+            </label>
+            <select
+              value={formData.work_schedule_id}
+              onChange={(e) => setFormData({ ...formData, work_schedule_id: e.target.value })}
+              className="w-full p-2 border border-gray-300 rounded-md focus:ring-2 focus:ring-blue-500"
+            >
+              <option value="">Select Schedule</option>
+              {/* eslint-disable-next-line react/jsx-key */}
+              {workSchedules.map((schedule, index) => (
+                <option key={`schedule-${index}`} value={schedule.id}>{schedule.name}</option>
+              ))}
+            </select>
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-gray-700 mb-1">
+              Base Salary (HNL)
+            </label>
+            <Input
+              type="number"
+              step="0.01"
+              value={formData.base_salary}
+              onChange={(e) => setFormData({ ...formData, base_salary: e.target.value })}
+              placeholder="25000.00"
+              required
+            />
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-gray-700 mb-1">
+              Hire Date
+            </label>
+            <Input
+              type="date"
+              value={formData.hire_date}
+              onChange={(e) => setFormData({ ...formData, hire_date: e.target.value })}
+              required
+            />
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-gray-700 mb-1">
+              Bank Name
+            </label>
+            <Input
+              value={formData.bank_name}
+              onChange={(e) => setFormData({ ...formData, bank_name: e.target.value })}
+              placeholder="Banco Atlántida"
+            />
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-gray-700 mb-1">
+              Bank Account
+            </label>
+            <Input
+              value={formData.bank_account}
+              onChange={(e) => setFormData({ ...formData, bank_account: e.target.value })}
+              placeholder="12345678901"
+            />
+          </div>
+
+          <div className="md:col-span-2 flex gap-4">
+            <Button type="submit" disabled={loading}>
+              {loading ? 'Adding...' : 'Add Employee'}
+            </Button>
+            <Button
+              type="button"
+              variant="outline"
+              onClick={onCancel}
+            >
+              Cancel
+            </Button>
+          </div>
+        </form>
+      </CardContent>
+    </Card>
+  )
+}
